Updated disease_ontology.yaml
I have updated the ontology files adding a couple fo diseases that are in the package and were missing. Some term cannot reliably be mapped to an ontology mainly because i) is not of super-wide adoption (SRP: scale root planing (therapy for periodontitis)) ii) because the relative knowledge is still small (e.g. FMT, fecal metagenome transplantation). Other terms are not mappable because they do not correspond to a well-defined condition but rather to a condition defined by the study observations (T1D non-converter are healthy individuals which differ from random people because they did not show sero-conversion). 
All the other terms seem to be mapped correctly right now.